chore: refactor web archi
