+++ demo/web/src/assets/game-controller.png
