+++ demo/web/src/css/App.css
@@ -17,7 +17,7 @@
 }
 
 .App-header {
-  background-image: url("./../public/background.png");
+  background-image: url("../../public/background.png");
   min-height: 100vh;
   display: flex;
   flex-direction: column;
